+++ spm-monitor-generic/src/main/java/com/sematext/spm/client/functions/DoubleTrimTimeUnit.java
@@ -2,8 +2,8 @@ package com.sematext.spm.client.functions;
 
 import java.util.concurrent.TimeUnit;
 
-public class DoubleTrimTimeUnit extends LongTrimTimeUnit {
-    protected Object convert(String trimmedValue, TimeUnit to, TimeUnit from) {
+public class DoubleTrimTimeUnit extends TrimTimeUnit {
+    protected Number convert(String trimmedValue, TimeUnit to, TimeUnit from) {
         double time = Double.parseDouble(trimmedValue);
         if (from.ordinal() < to.ordinal()) {
             return time / from.convert(1, to);
